refactor: move components, replace svg placeholder with background #136
