+++ src/features/Search/pages/CharacterSearchPage/components/CharacterCoverCard/CharacterCoverCard.tsx
@@ -1,7 +1,6 @@
 import React, { memo, VFC } from 'react';
-import { CoverCard } from '../CoverCard/CoverCard';
-import { Character } from '../../../../models/characters/character';
-import imagePlaceholder from '../../../../assets/person.svg';
+import { CoverCard } from '../../../../components/CoverCard/CoverCard';
+import { Character } from '../../../../../../models/characters/character';
 
 interface Props {
 
@@ -28,7 +27,6 @@ const CharacterCoverCardComponent: VFC<Props> = ({
     title={name}
     image={image}
     isImageNsfw={isImageNsfw}
-    imagePlaceholder={imagePlaceholder}
   />
 );
