+++ src/features/Search/pages/VisualNovelSearchPage/components/VisualNovelList/VisualNovelList.tsx
@@ -1,15 +1,15 @@
 import React, { memo, VFC } from 'react';
 import { Box } from '@chakra-ui/react';
-import { VisualNovel } from '../../../../models/visualNovels/visualNovel';
-import { CoverCardSkeleton } from '../CoverCard/CoverCardSkeleton';
+import { VisualNovel } from '../../../../../../models/visualNovels/visualNovel';
+import { CoverCardSkeleton } from '../../../../components/CoverCard/CoverCardSkeleton';
 import { VisualNovelCoverCard } from '../VisualNovelCoverCard/VisualNovelCoverCard';
 import { VisualNovelTable } from '../VisualNovelTable/VisualNovelTable';
 import { VisualNovelTableSkeleton } from '../VisualNovelTable/VisualNovelTableSkeleton';
-import { WideCard } from '../WideCard/WideCard';
-import { WideCardSkeleton } from '../WideCard/WideCardSkeleton';
-import { ExtendedCard } from '../ExtendedCard/ExtendedCard';
-import { ExtendedCardSkeleton } from '../ExtendedCard/ExtendedCardSkeleton';
-import { checkImageNsfw } from '../../../../utils/checkImageNsfw';
+import { WideCard } from '../../../../components/WideCard/WideCard';
+import { WideCardSkeleton } from '../../../../components/WideCard/WideCardSkeleton';
+import { ExtendedCard } from '../../../../components/ExtendedCard/ExtendedCard';
+import { ExtendedCardSkeleton } from '../../../../components/ExtendedCard/ExtendedCardSkeleton';
+import { checkImageNsfw } from '../../../../../../utils/checkImageNsfw';
 
 export type VisualNovelListVariant = 'table' | 'cards' | 'wide-cards' | 'extended-cards';
