+++ src/features/Search/pages/VisualNovelSearchPage/components/VisualNovelListOptions/VisualNovelListOptions.tsx
@@ -1,6 +1,6 @@
 import React, { memo, useCallback, VFC } from 'react';
 import { Box, IconButton } from '@chakra-ui/react';
-import { Icon } from '../../../../components/Icon/Icon';
+import { Icon } from '../../../../../../components/Icon/Icon';
 import { VisualNovelListVariant } from '../VisualNovelList/VisualNovelList';
 
 interface Props {
